+++ elasticsearch/common/src/test/java/org/apache/metamodel/elasticsearch/common/ElasticSearchUtilsTest.java
@@ -18,15 +18,26 @@
  */
 package org.apache.metamodel.elasticsearch.common;
 
-import junit.framework.TestCase;
+import java.util.Arrays;
+import java.util.Collections;
+import java.util.Date;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
+
 import org.apache.metamodel.data.DataSetHeader;
 import org.apache.metamodel.data.Row;
 import org.apache.metamodel.data.SimpleDataSetHeader;
+import org.apache.metamodel.query.FilterItem;
+import org.apache.metamodel.query.OperatorType;
 import org.apache.metamodel.query.SelectItem;
 import org.apache.metamodel.schema.ColumnType;
 import org.apache.metamodel.schema.MutableColumn;
+import org.elasticsearch.common.xcontent.XContentHelper;
+import org.elasticsearch.common.xcontent.XContentType;
+import org.elasticsearch.index.query.QueryBuilder;
 
-import java.util.*;
+import junit.framework.TestCase;
 
 public class ElasticSearchUtilsTest extends TestCase {
 
@@ -59,5 +70,64 @@ public class ElasticSearchUtilsTest extends TestCase {
 
         assertTrue(stringValue instanceof String);
         assertTrue(dateValue instanceof Date);
+    }
+
+    /**
+     * For text-based conditions a 'match' query is recommended (instead of 'term' query).
+     */
+    public void testMatchQueryIsCreatedForTextEqualTo() {
+        final SelectItem selectItem = new SelectItem(new MutableColumn("column_name", ColumnType.STRING));
+        final FilterItem filterItem = new FilterItem(selectItem, OperatorType.EQUALS_TO, "text-value");
+        final QueryBuilder queryBuilder =
+                ElasticSearchUtils.createQueryBuilderForSimpleWhere(Collections.singletonList(filterItem), null);
+        assertNotNull(queryBuilder);
+        assertEquals("match", queryBuilder.getName());
+    }
+
+    /**
+     * For text-based conditions a 'match' query is recommended (instead of 'term' query).
+     * In case of 'DIFFERENT_FROM', we need a 'bool' query with 'must not' and 'match' query.
+     */
+    @SuppressWarnings("unchecked")
+    public void testBoolQueryIsCreatedForTextDifferentFrom() {
+        final SelectItem selectItem = new SelectItem(new MutableColumn("column_name", ColumnType.STRING));
+        final FilterItem filterItem = new FilterItem(selectItem, OperatorType.DIFFERENT_FROM, "text-value");
+        final QueryBuilder queryBuilder =
+                ElasticSearchUtils.createQueryBuilderForSimpleWhere(Collections.singletonList(filterItem), null);
+        assertNotNull(queryBuilder);
+        assertEquals("bool", queryBuilder.getName());
+        final Map<String, Object> queryMap =
+                XContentHelper.convertToMap(XContentType.JSON.xContent(), queryBuilder.toString(), false);
+        final Map<String, Object> boolMap = (Map<String, Object>) queryMap.get("bool");
+        assertNotNull(boolMap);
+        assertTrue(boolMap.containsKey("must_not"));
+        final List<Object> mustNotList = (List<Object>) boolMap.get("must_not");
+        final Map<String, Object> mustNotMap = (Map<String, Object>) mustNotList.get(0);
+        assertTrue(mustNotMap.containsKey("match"));
+    }
+
+    /**
+     * For text-based conditions a 'match' query is recommended (instead of 'term' query).
+     * To simulate 'IN' operator, we need a 'bool' query with multiple 'match' queries combined with 'OR'.
+     */
+    @SuppressWarnings("unchecked")
+    public void testBoolQueryIsCreatedForTextIn() {
+        final SelectItem selectItem = new SelectItem(new MutableColumn("column_name", ColumnType.STRING));
+        final FilterItem filterItem =
+                new FilterItem(selectItem, OperatorType.IN, Arrays.asList("text-value-a", "text-value-b"));
+        final QueryBuilder queryBuilder =
+                ElasticSearchUtils.createQueryBuilderForSimpleWhere(Collections.singletonList(filterItem), null);
+        assertNotNull(queryBuilder);
+        assertEquals("bool", queryBuilder.getName());
+        final Map<String, Object> queryMap =
+                XContentHelper.convertToMap(XContentType.JSON.xContent(), queryBuilder.toString(), false);
+        final Map<String, Object> matchMap = (Map<String, Object>) queryMap.get("bool");
+        assertNotNull(matchMap);
+        assertTrue(matchMap.containsKey("should"));
+        final List<Object> shouldList = (List<Object>) matchMap.get("should");
+        assertNotNull(shouldList);
+        assertEquals(2, shouldList.size());
+        assertTrue(((Map<String, Object>)shouldList.get(0)).containsKey("match"));
+        assertTrue(((Map<String, Object>)shouldList.get(1)).containsKey("match"));
     }
 }