+++ jdbc/src/test/java/org/apache/metamodel/jdbc/integrationtests/FirebirdTest.java
@@ -18,14 +18,10 @@
  */
 package org.apache.metamodel.jdbc.integrationtests;
 
-import java.text.SimpleDateFormat;
-import java.util.Arrays;
-import java.util.Date;
-import java.util.Set;
-import java.util.stream.Collectors;
+import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertNotNull;
 
-import javax.swing.table.TableModel;
-
+import com.google.common.collect.Sets;
 import org.apache.metamodel.data.DataSet;
 import org.apache.metamodel.data.DataSetTableModel;
 import org.apache.metamodel.jdbc.JdbcDataContext;
@@ -35,8 +31,14 @@ import org.apache.metamodel.query.Query;
 import org.apache.metamodel.query.SelectItem;
 import org.apache.metamodel.schema.Schema;
 import org.apache.metamodel.schema.Table;
+import org.junit.Test;
 
-import com.google.common.collect.Sets;
+import javax.swing.table.TableModel;
+import java.text.SimpleDateFormat;
+import java.util.Arrays;
+import java.util.Date;
+import java.util.Set;
+import java.util.stream.Collectors;
 
 /**
  * Integrationtests for Firebird SQL.
@@ -55,6 +57,7 @@ public class FirebirdTest extends AbstractJdbIntegrationTest {
         return "firebird";
     }
 
+    @Test
 	public void testGetSchemas() throws Exception {
 	    if (!isConfigured()) {
 	        return;
@@ -97,6 +100,7 @@ public class FirebirdTest extends AbstractJdbIntegrationTest {
 		assertEquals(relationStrings,referenceRelationStrings);
 	}
 
+	@Test
 	public void testExecuteQuery() throws Exception {
 	    if (!isConfigured()) {
             return;