+++ jdbc/src/test/java/org/apache/metamodel/jdbc/integrationtests/HiveIntegrationTest.java
@@ -18,6 +18,8 @@
  */
 package org.apache.metamodel.jdbc.integrationtests;
 
+import static org.junit.Assert.*;
+
 import java.sql.Connection;
 import java.sql.SQLException;
 import java.sql.Statement;
@@ -34,6 +36,8 @@ import org.apache.metamodel.schema.Schema;
 import org.apache.metamodel.schema.Table;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+
+import org.junit.Test;
 
 public class HiveIntegrationTest extends AbstractJdbIntegrationTest {
     private static final Logger logger = LoggerFactory.getLogger(HiveIntegrationTest.class);
@@ -61,6 +65,7 @@ public class HiveIntegrationTest extends AbstractJdbIntegrationTest {
         logger.info("SQL updated fired (return {}): {}", st.executeUpdate(deleteLastSql), deleteLastSql);
     }
 
+    @Test
     public void testDefaultGetSchema() throws Exception {
         if (!isConfigured()) {
             return;
@@ -87,6 +92,7 @@ public class HiveIntegrationTest extends AbstractJdbIntegrationTest {
         }
     }
 
+    @Test
     public void testGetSchema() throws Exception {
         if (!isConfigured()) {
             return;
@@ -97,6 +103,7 @@ public class HiveIntegrationTest extends AbstractJdbIntegrationTest {
         assertEquals("Schema[name=default]", schema.toString());
     }
 
+    @Test
     public void testUseCorrectRewriter() throws Exception {
         if (!isConfigured()) {
             return;
@@ -106,6 +113,7 @@ public class HiveIntegrationTest extends AbstractJdbIntegrationTest {
         assertTrue(dataContext.getQueryRewriter() instanceof HiveQueryRewriter);
     }
 
+    @Test
     public void testCreateInsertQueryAndDrop() throws Exception {
         if (!isConfigured()) {
             return;
@@ -155,6 +163,7 @@ public class HiveIntegrationTest extends AbstractJdbIntegrationTest {
      * MetaModel should ignore that for Hive datastores when creating a table.
      * @throws Exception
      */
+    @Test
     public void testCreateTableWithPrimaryKey() throws Exception {
         if (!isConfigured()) {
             return;