+++ jdbc/src/test/java/org/apache/metamodel/jdbc/integrationtests/MysqlTest.java
@@ -18,6 +18,8 @@
  */
 package org.apache.metamodel.jdbc.integrationtests;
 
+import static org.junit.Assert.*;
+
 import java.sql.Connection;
 import java.sql.DatabaseMetaData;
 import java.sql.SQLException;
@@ -50,6 +52,7 @@ import org.apache.metamodel.schema.ColumnType;
 import org.apache.metamodel.schema.Schema;
 import org.apache.metamodel.schema.Table;
 import org.apache.metamodel.schema.TableType;
+import org.junit.Test;
 
 /**
  * Test case that tests mysql interaction. The test requires the "sakila" sample
@@ -64,6 +67,7 @@ public class MysqlTest extends AbstractJdbIntegrationTest {
         return "mysql";
     }
 
+    @Test
     public void testCreateInsertAndUpdate() throws Exception {
         if (!isConfigured()) {
             return;
@@ -72,6 +76,7 @@ public class MysqlTest extends AbstractJdbIntegrationTest {
         JdbcTestTemplates.simpleCreateInsertUpdateAndDrop(getDataContext(), "metamodel_test_simple");
     }
 
+    @Test
     public void testCompositePrimaryKeyCreation() throws Exception {
         if (!isConfigured()) {
             return;
@@ -80,6 +85,7 @@ public class MysqlTest extends AbstractJdbIntegrationTest {
         JdbcTestTemplates.compositeKeyCreation(getDataContext(), "metamodel_test_composite_keys");
     }
 
+    @Test
     public void testTimestampValueInsertSelect() throws Exception {
         if (!isConfigured()) {
             return;
@@ -89,6 +95,7 @@ public class MysqlTest extends AbstractJdbIntegrationTest {
         JdbcTestTemplates.timestampValueInsertSelect(connection, TimeUnit.MICROSECONDS, "TIMESTAMP(6)");
     }
 
+    @Test
     public void testInterpretationOfNull() throws Exception {
         if (!isConfigured()) {
             return;
@@ -97,6 +104,7 @@ public class MysqlTest extends AbstractJdbIntegrationTest {
         JdbcTestTemplates.interpretationOfNulls(getConnection());
     }
 
+    @Test
     public void testDatabaseProductName() throws Exception {
         if (!isConfigured()) {
             return;
@@ -106,6 +114,7 @@ public class MysqlTest extends AbstractJdbIntegrationTest {
         assertEquals(JdbcDataContext.DATABASE_PRODUCT_MYSQL, databaseProductName);
     }
 
+    @Test
     public void testCreateInsertAndUpdateDateTypes() throws Exception {
         if (!isConfigured()) {
             return;
@@ -116,6 +125,7 @@ public class MysqlTest extends AbstractJdbIntegrationTest {
                 "metamodel_mysql_test");
     }
 
+    @Test
     public void testAutomaticConversionWhenInsertingString() throws Exception {
         if (!isConfigured()) {
             return;
@@ -170,6 +180,7 @@ public class MysqlTest extends AbstractJdbIntegrationTest {
         });
     }
 
+    @Test
     public void testCharOfSizeOne() throws Exception {
         if (!isConfigured()) {
             return;
@@ -178,6 +189,7 @@ public class MysqlTest extends AbstractJdbIntegrationTest {
         JdbcTestTemplates.meaningOfOneSizeChar(getConnection());
     }
 
+    @Test
     public void testAlternativeConnectionString() throws Exception {
         if (!isConfigured()) {
             return;
@@ -195,6 +207,7 @@ public class MysqlTest extends AbstractJdbIntegrationTest {
         assertEquals(1000, tableModel.getRowCount());
     }
 
+    @Test
     public void testGetCatalogNames() throws Exception {
         if (!isConfigured()) {
             return;
@@ -207,6 +220,7 @@ public class MysqlTest extends AbstractJdbIntegrationTest {
         assertNotNull(dataContext.getSchemaByName("sakila"));
     }
 
+    @Test
     public void testGetDefaultSchema() throws Exception {
         if (!isConfigured()) {
             return;
@@ -217,6 +231,7 @@ public class MysqlTest extends AbstractJdbIntegrationTest {
         assertEquals("sakila", schema.getName());
     }
 
+    @Test
     public void testExecuteQuery() throws Exception {
         if (!isConfigured()) {
             return;
@@ -282,6 +297,7 @@ public class MysqlTest extends AbstractJdbIntegrationTest {
     }
 
     // Test to query the film table (caused troubles in DataCleaner)
+    @Test
     public void testFilmQuery() throws Exception {
         if (!isConfigured()) {
             return;
@@ -293,6 +309,7 @@ public class MysqlTest extends AbstractJdbIntegrationTest {
         dc.executeQuery(q);
     }
 
+    @Test
     public void testGetSchema() throws Exception {
         if (!isConfigured()) {
             return;
@@ -364,6 +381,7 @@ public class MysqlTest extends AbstractJdbIntegrationTest {
                 Arrays.toString(staffView.getColumns().toArray()));
     }
 
+    @Test
     public void testSplitQuery() throws Exception {
         if (!isConfigured()) {
             return;
@@ -400,6 +418,7 @@ public class MysqlTest extends AbstractJdbIntegrationTest {
         assertEquals(32098, count);
     }
 
+    @Test
     public void testQueryWithSingleQuote() throws Exception {
         if (!isConfigured()) {
             return;
@@ -413,6 +432,7 @@ public class MysqlTest extends AbstractJdbIntegrationTest {
         assertFalse(ds.next());
     }
 
+    @Test
     public void testCompiledQueries() throws Exception {
         if (!isConfigured()) {
             return;
@@ -444,6 +464,7 @@ public class MysqlTest extends AbstractJdbIntegrationTest {
         }
     }
 
+    @Test
     public void testWhiteSpaceColumns() throws Exception {
         if (!isConfigured()) {
             return;