+++ jdbc/src/test/java/org/apache/metamodel/jdbc/integrationtests/SQLServerJtdsDriverTest.java
@@ -18,6 +18,8 @@
  */
 package org.apache.metamodel.jdbc.integrationtests;
 
+import static org.junit.Assert.*;
+
 import java.sql.Connection;
 import java.sql.SQLException;
 import java.util.Arrays;
@@ -41,6 +43,8 @@ import org.apache.metamodel.schema.Schema;
 import org.apache.metamodel.schema.Table;
 import org.apache.metamodel.schema.TableType;
 
+import org.junit.Test;
+
 /**
  * Test case that tests MS SQL Server interaction. The test uses the
  * "AdventureWorks 2012" sample database which can be downloaded from codeplex.
@@ -58,6 +62,7 @@ public class SQLServerJtdsDriverTest extends AbstractJdbIntegrationTest {
         return "sqlserver.jtds_driver";
     }
 
+    @Test
     public void testCreateInsertAndUpdate() throws Exception {
         if (!isConfigured()) {
             return;
@@ -66,6 +71,7 @@ public class SQLServerJtdsDriverTest extends AbstractJdbIntegrationTest {
     }
 
     // This test is pretty useless. It assumes way too much, and fails due to SQL Server not using timestamp as assumed.
+    @Test
     public void ignoreTestTimestampValueInsertSelect() throws Exception {
         if (!isConfigured()) {
             return;
@@ -75,6 +81,7 @@ public class SQLServerJtdsDriverTest extends AbstractJdbIntegrationTest {
         JdbcTestTemplates.timestampValueInsertSelect(connection, TimeUnit.NANOSECONDS, "datetime");
     }
 
+    @Test
     public void testCreateTableInUpdateScript() throws Exception {
         if (!isConfigured()) {
             return;
@@ -104,6 +111,7 @@ public class SQLServerJtdsDriverTest extends AbstractJdbIntegrationTest {
         dc.executeUpdate(new DropTable(schema, tableName));
     }
 
+    @Test
     public void testCompositePrimaryKeyCreation() throws Exception {
         if (!isConfigured()) {
             return;
@@ -112,6 +120,7 @@ public class SQLServerJtdsDriverTest extends AbstractJdbIntegrationTest {
         JdbcTestTemplates.compositeKeyCreation(getDataContext(), "metamodel_test_composite_keys");
     }
 
+    @Test
     public void testWorkingWithDates() throws Exception {
         if (!isConfigured()) {
             return;
@@ -125,6 +134,7 @@ public class SQLServerJtdsDriverTest extends AbstractJdbIntegrationTest {
         JdbcTestTemplates.createInsertAndUpdateDateTypes(dc, schema, "test_table");
     }
 
+    @Test
     public void testAutomaticConversionWhenInsertingString() throws Exception {
         if (!isConfigured()) {
             return;
@@ -179,6 +189,7 @@ public class SQLServerJtdsDriverTest extends AbstractJdbIntegrationTest {
         connection.createStatement().execute("DROP TABLE Person.test_table");
     }
 
+    @Test
     public void testQueryUsingExpressions() throws Exception {
         if (!isConfigured()) {
             return;
@@ -199,6 +210,7 @@ public class SQLServerJtdsDriverTest extends AbstractJdbIntegrationTest {
         assertFalse(dataSet.next());
     }
 
+    @Test
     public void testGetSchemaNormalTableTypes() throws Exception {
         if (!isConfigured()) {
             return;
@@ -234,6 +246,7 @@ public class SQLServerJtdsDriverTest extends AbstractJdbIntegrationTest {
         assertEquals(26, salesSchema.getTableCount());
     }
 
+    @Test
     public void testGetSchemaAllTableTypes() throws Exception {
         if (!isConfigured()) {
             return;
@@ -249,6 +262,7 @@ public class SQLServerJtdsDriverTest extends AbstractJdbIntegrationTest {
         assertTrue(strategy.getSchemaNames().containsAll(expectedSchemaNames));
     }
 
+    @Test
     public void testQueryRewriterQuoteAliases() throws Exception {
         if (!isConfigured()) {
             return;
@@ -288,6 +302,7 @@ public class SQLServerJtdsDriverTest extends AbstractJdbIntegrationTest {
         data.close();
     }
 
+    @Test
     public void testQuotedString() throws Exception {
         if (!isConfigured()) {
             return;
@@ -308,6 +323,7 @@ public class SQLServerJtdsDriverTest extends AbstractJdbIntegrationTest {
                 queryRewriter.rewriteQuery(q));
     }
 
+    @Test
     public void testMaxAndOffset() throws Exception {
         if (!isConfigured()) {
             return;