+++ jdbc/src/test/java/org/apache/metamodel/jdbc/integrationtests/SQLServerMicrosoftDriverTest.java
@@ -18,6 +18,8 @@
  */
 package org.apache.metamodel.jdbc.integrationtests;
 
+import static org.junit.Assert.*;
+
 import java.util.Arrays;
 
 import org.apache.metamodel.data.DataSet;
@@ -29,6 +31,8 @@ import org.apache.metamodel.query.SelectItem;
 import org.apache.metamodel.schema.Schema;
 import org.apache.metamodel.schema.Table;
 import org.apache.metamodel.schema.TableType;
+
+import org.junit.Test;
 
 /**
  * Test case that tests MS SQL Server interaction. The test uses the
@@ -47,6 +51,7 @@ public class SQLServerMicrosoftDriverTest extends AbstractJdbIntegrationTest {
         return "sqlserver.microsoft_driver";
     }
 
+    @Test
     public void testQueryUsingExpressions() throws Exception {
         if (!isConfigured()) {
             return;
@@ -66,6 +71,7 @@ public class SQLServerMicrosoftDriverTest extends AbstractJdbIntegrationTest {
         assertFalse(dataSet.next());
     }
 
+    @Test
     public void testGetSchemaNormalTableTypes() throws Exception {
         if (!isConfigured()) {
             return;
@@ -91,6 +97,7 @@ public class SQLServerMicrosoftDriverTest extends AbstractJdbIntegrationTest {
 
     }
 
+    @Test
     public void testGetSchemaAllTableTypes() throws Exception {
         if (!isConfigured()) {
             return;
@@ -105,6 +112,7 @@ public class SQLServerMicrosoftDriverTest extends AbstractJdbIntegrationTest {
         assertEquals("Schema[name=dbo]", strategy.getDefaultSchema().toString());
     }
 
+    @Test
     public void testQueryRewriterQuoteAliases() throws Exception {
         if (!isConfigured()) {
             return;
@@ -145,6 +153,7 @@ public class SQLServerMicrosoftDriverTest extends AbstractJdbIntegrationTest {
         data.close();
     }
 
+    @Test
     public void testQuotedString() throws Exception {
         if (!isConfigured()) {
             return;