+++ CHANGES.md
@@ -1,5 +1,6 @@
-### Apache MetaModel [WIP]
+### Apache MetaModel 5.3.3
 
+ * [METAMODEL-1233] - Elasticsearch: "term" query is used for text search.
  * [METAMODEL-1229] - FixedWidth works with UrlResource/HttpInputStream.
  * [METAMODEL-1228] - Better handling of fieldnames with dots in Elasticsearch
  * [METAMODEL-1227] - Better handling of nested objects in Elasticsearch data